Buildable version, untested

+++ src/network.c
@@ -175,7 +175,7 @@ network make_network(int n)
     net.n = n;
 
     net.layers = (layer*)calloc(net.n, sizeof(layer));
-    net.seen = (int*)calloc(1, sizeof(int));
+    net.seen = (size_t*)calloc(1, sizeof(size_t));
     net.t    = (int*)calloc(1, sizeof(int));
     net.cost = (float*)calloc(1, sizeof(float));
     return net;
